DELETE method button not generated correctly
Non GET type buttons are generated as links on line 45 but remain in the config array. 
This causes a fatal error on line 74 when ['options]['class'] is set on a string. If the link
was already generated then we can just echo that string and continue building the rest
of the links from the config.